server/requirements.txt
google-genai
python-dotenv
uvicorn
fastapi
websockets
twilio
pydantic
